+++ src/eop_data_iers_rapid.cpp
@@ -188,9 +188,6 @@ eop_data_table parse_eop_data_iers_rapid(const std::string &str)
             {.mjd = mjd, .delta_ut1_utc = *delta_ut1_utc, .pm_x = *pm_x, .pm_y = *pm_y, .dX = *dX, .dY = *dY});
     }
 
-    // Validate the output.
-    validate_eop_data_table(retval);
-
     return retval;
 }
 
@@ -256,8 +253,8 @@ eop_data eop_data::fetch_latest_iers_rapid(const std::string &origin, const std:
     // primitives (which use '.' as a separator).
     std::ranges::replace(identifier, '.', '_');
 
-    // Parse, validate and return.
-    return eop_data(detail::parse_eop_data_iers_rapid(text), std::move(timestamp), std::move(identifier));
+    // Parse, construct and return.
+    return eop_data(detail::parse_eop_data_iers_rapid(text), std::move(timestamp), std::move(identifier), true);
 }
 
 HEYOKA_END_NAMESPACE